refactor: move CAddressBookData to addressbookman.h
