+++ src/Makefile.am
@@ -298,6 +298,7 @@ BITCOIN_CORE_H = \
   validation.h \
   validationinterface.h \
   versionbits.h \
+  wallet/addressbookman.h \
   wallet/bdb.h \
   wallet/coincontrol.h \
   wallet/coinselection.h \
