+++ src/wallet/addressbookman.h
@@ -1,0 +1,30 @@
+// Copyright (c) 2022 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_WALLET_ADDRESSBOOKMAN_H
+#define BITCOIN_WALLET_ADDRESSBOOKMAN_H
+
+/** Address book data */
+class CAddressBookData
+{
+private:
+    bool m_change{true};
+    std::string m_label;
+public:
+    std::string purpose;
+
+    CAddressBookData() : purpose("unknown") {}
+
+    typedef std::map<std::string, std::string> StringMap;
+    StringMap destdata;
+
+    bool IsChange() const { return m_change; }
+    const std::string& GetLabel() const { return m_label; }
+    void SetLabel(const std::string& label) {
+        m_change = false;
+        m_label = label;
+    }
+};
+
+#endif // BITCOIN_WALLET_ADDRESSBOOKMAN_H
